backend/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
celery>=5.3
redis>=5.0
django-cors-headers>=4.3
